+++ modules/HardwareDrivers/PowerSupplies/Huawei_V100R023C10/fusion_charger_lib/fusion-charger-dispenser-library/lib/connector.cpp
@@ -160,8 +160,9 @@ std::string state_to_string(States state) {
     return "UNKNOWN";
 }
 
-Connector::Connector(ConnectorConfig connector_config, std::uint16_t local_connector_number,
-                     DispenserConfig dispenser_config, logs::LogIntf log) :
+Connector::Connector(ConnectorConfig connector_config, uint16_t local_connector_number,
+                     DispenserConfig dispenser_config, logs::LogIntf log,
+                     std::function<void()> do_unsolicitated_report_callback) :
     connector_config(connector_config),
     local_connector_number(local_connector_number),
     dispenser_config(dispenser_config),
@@ -187,6 +188,7 @@ Connector::Connector(ConnectorConfig connector_config, std::uint16_t local_conne
         config.get_output_current = connector_config.connector_callbacks.output_current;
         config.get_contactor_status = connector_config.connector_callbacks.contactor_status;
         config.get_electronic_lock_status = connector_config.connector_callbacks.electronic_lock_status;
+        config.get_dc_output_contact_fault = [this]() { return dc_output_contactor_fault_alarm_active.load(); };
         return config;
     }()),
     connector_registers(connector_registers_config),
@@ -199,7 +201,8 @@ Connector::Connector(ConnectorConfig connector_config, std::uint16_t local_conne
                                                  std::placeholders::_1, std::placeholders::_2);
             return callbacks;
         }(),
-        log, log_prefix) {
+        log, log_prefix),
+    do_unsolicitated_report_callback(do_unsolicitated_report_callback) {
 }
 
 Connector::~Connector() {
@@ -314,6 +317,26 @@ void Connector::start() {
     // todo: reset fsm?
 
     goose_sender.start();
+
+    std::string connector_telemetry_subtopic = "connector/" + std::to_string(connector_config.global_connector_number);
+    dispenser_config.telemetry_manager->add_subtopic(connector_telemetry_subtopic);
+
+    dispenser_config.telemetry_manager->register_complex_register_data_provider(
+        connector_telemetry_subtopic, "max_rated_psu_current", &connector_registers.max_rated_psu_current);
+    dispenser_config.telemetry_manager->register_complex_register_data_provider(
+        connector_telemetry_subtopic, "min_rated_psu_current", &connector_registers.min_rated_psu_current);
+    dispenser_config.telemetry_manager->register_complex_register_data_provider(
+        connector_telemetry_subtopic, "max_rated_psu_voltage", &connector_registers.max_rated_psu_voltage);
+    dispenser_config.telemetry_manager->register_complex_register_data_provider(
+        connector_telemetry_subtopic, "min_rated_psu_voltage", &connector_registers.min_rated_psu_voltage);
+
+    dispenser_config.telemetry_manager->register_complex_register_data_provider<float>(
+        connector_telemetry_subtopic, "rated_output_power_psu", &connector_registers.rated_output_power_psu,
+        [](const float& kw) { return kw * 1000.0; });
+
+    dispenser_config.telemetry_manager->register_complex_register_data_provider_enum<PsuOutputPortAvailability>(
+        connector_telemetry_subtopic, "psu_port_available", &connector_registers.psu_port_available,
+        psu_output_port_availability_to_string);
 }
 
 void Connector::stop() {
@@ -524,4 +547,13 @@ void Connector::on_psu_mac_change(std::vector<std::uint8_t> mac_address) {
 std::vector<std::uint8_t> Connector::get_hmac_key() {
     const std::uint8_t* hmac_key = connector_registers.hmac_key.get_value(); // pointer to private memory
     return std::vector<std::uint8_t>(hmac_key, hmac_key + connector_registers.hmac_key.get_size());
+}
+
+void Connector::set_dc_output_contactor_fault_alarm(bool active) {
+    dc_output_contactor_fault_alarm_active = active;
+
+    // immediately do an unsolicitated report
+    if (do_unsolicitated_report_callback) {
+        do_unsolicitated_report_callback();
+    }
 }